apps/api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy[asyncio]==2.0.31
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic-settings==2.3.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-multipart==0.0.9
redis[hiredis]==5.0.7
pinecone-client==4.0.0
langchain==0.2.6
langchain-openai==0.1.14
langchain-community==0.2.6
openai==1.35.7
httpx==0.27.0
python-dotenv==1.0.1
PyPDF2==3.0.1
python-docx==1.2.0
